Trait libsmt::backends::backend::SMTBackend

pub trait SMTBackend {
    type Idx: Debug + Clone;
    type Logic: Logic;
    fn set_logic<S: SMTProc>(&mut self, &mut S);
    fn new_var<T, P>(&mut self, Option<T>, P) -> Self::Idx where T: AsRef<str>, P: Into<Self::Logic::Sorts>;
    fn assert<T: Into<Self::Logic::Fns>>(&mut self, T, &[Self::Idx]) -> Self::Idx;
    fn check_sat<S: SMTProc>(&mut self, &mut S) -> bool;
    fn solve<S: SMTProc>(&mut self, &mut S) -> SMTResult<HashMap<Self::Idx, u64>>;
}

Trait a backend should implement to support SMT solving.

This is a minimalistic API and has to be expanded in the future to support more SMT operations and to grow this into a full SMTLib Crate.

All functions names are analogous in meaning to their usage in the original SMT-LIB2 sense. TODO: - define_fun - declare_sort - define_sort - get_proof - get_unsat_core - get_value - get_assignment - push - pop - get_option - set_option - get_info - set_info - exit

Associated Types

type Idx: Debug + Clone

type Logic: Logic

Required Methods

fn set_logic<S: SMTProc>(&mut self, &mut S)

fn new_var<T, P>(&mut self, Option<T>, P) -> Self::Idx where T: AsRef<str>, P: Into<Self::Logic::Sorts>

fn assert<T: Into<Self::Logic::Fns>>(&mut self, T, &[Self::Idx]) -> Self::Idx

fn check_sat<S: SMTProc>(&mut self, &mut S) -> bool

fn solve<S: SMTProc>(&mut self, &mut S) -> SMTResult<HashMap<Self::Idx, u64>>