Trait libsmt::backends::backend::SMTNode

pub trait SMTNode: Display {
    fn is_var(&self) -> bool;
    fn is_const(&self) -> bool;

    fn is_fn(&self) -> bool { ... }
}

Required Methods

fn is_var(&self) -> bool

Returns true if the node is a symbolic variable

fn is_const(&self) -> bool

Returns true if the node is a constant value

Provided Methods

fn is_fn(&self) -> bool

Returns true if the node is a function