Module libsmt::theories

Modules

array_ex
Module that describes the ArrayEx Theory

bitvec
Defines basic operations defined under FixedSizeBitVectors theory in SMTLIB2.

core
Defines basic operation defined under Core theory in SMTLIB2.

integer
Defines basic operations defined under Int theory in SMTLIB2.

real
Defines basic operations defined under Real theory in SMTLIB2.

real_ints
Defines basic operations defined under Real_Ints theory in SMTLIB2.

utils
Macro helpers for defining theories